+++ vTeam/vTeam/Grid/NSString+Grid.m
@@ -38,7 +38,7 @@
     
     unichar uc;
     
-    int length = [self length];
+    NSUInteger length = [self length];
     int s = 0;
     
     for(int i=0;i<length;i++){
