+++ vTeam/vTeam/VTDataOutlet.m
@@ -73,7 +73,7 @@
     
     unichar uc;
     
-    int length = [self length];
+    NSUInteger length = [self length];
     int s = 0;
     
     for(int i=0;i<length;i++){
@@ -122,7 +122,7 @@
     
     unichar uc;
     
-    int length = [self length];
+    NSUInteger length = [self length];
     int s = 0;
     
     for(int i=0;i<length;i++){
@@ -172,7 +172,7 @@
     
     unichar uc;
     
-    int length = [self length];
+    NSUInteger length = [self length];
     int s = 0;
     
     for(int i=0;i<length;i++){
